Autodetect the baudrate for the Suunto D4i, D6i and D9tx.
These new models use a faster baudrate. The correct baudrate is detected
by probing with the version command. To speedup the detection, the model
code can be supplied by the caller as a hint, which is then used to
start the autodetection directly with the correct baudrate.

To maintain backwards compatibility, the model hint is not enabled yet.