Add support for a second Sherwood Insight variant.
There appears to be two very different versions of the Sherwood Insight.
The old Insight needs the veo250 backend, while the newer Insight 2
needs the atom2 backend. Currently only the newer version was included
in the list of supported devices, and to increase the confusion it was
even named after the old version.

With this patch, the old version is added to the list, and the new
version is renamed to "Insight 2".